Update hashes of DLed libs.
brainsik used sha1 for some reason but I switched to using sha256 since it is
already available in the PyPI page of the packages and is easily verifiable by
anyone who would care to verify.

Although at this point, partly due to my (justifiable!) laziness, we are really
dropping the ball on Python 2.